feat(recycle_bin, categories, audit, period_task, login): 回收站目录项：新增删除，还原功能
回收站内删除等同对象硬删除（相关资源一并删除），还原为对象恢复正常状态，操作计入审计日志；阻止异常状态目录用户登录；

BREAKING CHANGE: 原category.delete()将对应的资源disable且删除定时任务。新delete不会操作其资源，定期任务disable无效化,进行硬删除才将定时任务删除

feat #901

+++ src/api/bkuser_core/api/web/category/views.py
@@ -51,7 +51,7 @@ from bkuser_core.bkiam.permissions import (
     Permission,
     ViewCategoryPermission,
 )
-from bkuser_core.categories.constants import CategoryType, SyncTaskType
+from bkuser_core.categories.constants import CategoryStatus, CategoryType, SyncTaskType
 from bkuser_core.categories.exceptions import ExistsSyncingTaskError, FetchDataFromRemoteFailed
 from bkuser_core.categories.loader import get_plugin_by_category
 from bkuser_core.categories.models import ProfileCategory, SyncTask
@@ -289,12 +289,20 @@ class CategoryUpdateDeleteApi(generics.RetrieveUpdateDestroyAPIView):
     def delete(self, request, *args, **kwargs):
         """删除用户目录"""
         instance = self.get_object()
+
         if instance.default:
             raise error_codes.CANNOT_DELETE_DEFAULT_CATEGORY
 
+        if instance.status != CategoryStatus.INACTIVE.value:
+            raise error_codes.CANNOT_DELETE_ACTIVE_CATEGORY
+
         # 依赖 model 的 delete 方法, 执行软删除
         instance.delete()
-        post_category_delete.send_robust(sender=self, instance=instance, operator=request.operator)
+
+        # 善后：回收站映射，软删除审计日志
+        post_category_delete.send_robust(
+            sender=self, instance=instance, operator=request.operator, extra_values={"request": request}
+        )
         return Response(status=status.HTTP_200_OK)
